pyproject.toml
[project]
name = "dspy-workshop"
version = "0.2.0"
description = "Workshop demo: optimizing email parsing prompts with DSPy"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "dspy>=3.1.3",
    "notebook>=7.4.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.45.1",
]

[tool.ruff]
line-length = 120
